+++ src/components/Book.js
@@ -1,5 +1,10 @@
-const Book = (book = '') => {
-  <h1>{book}</h1>;
-};
+// eslint-disable-next-line react/prop-types
+const Book = ({ book: { ID, Title, category } }) => (
+  <tr>
+    <td>{ID}</td>
+    <td>{Title}</td>
+    <td>{category}</td>
+  </tr>
+);
 
 export default Book;
